requirements.txt
langchain
pdfminer.six
python-dotenv
streamlit
google.generativeai